取消PDOQuery
